Feat & Fix: Added in the new version of Arcade Game & fixed gulp-image issue
I have updated the arcade game within the portfolio, so that it included
the latest version of the game.

Gulp fix & Improvement
Fix: gulp-image stopped working after the version was updated to v2.0.1.
Had to manually set which image optimizations are set to ensure that
gulp-image would crash.

Feat: Gulp will now make sure mp3 audio files are moved to the 'dist'
folder as well.